+++ app/router.js
@@ -9,6 +9,7 @@ const Router = EmberRouter.extend({
     rootURL: config.rootURL,
     metrics: service(),
     theme: service(),
+    session: service(),
 
     didTransition() {
         this._super(...arguments);
@@ -17,10 +18,32 @@ const Router = EmberRouter.extend({
 
     _trackPage() {
         run.scheduleOnce('afterRender', this, () => {
+            // Tracks page with custom parameters
+            // authenticated => if the user is logged in or not
+            // isPublic      => if the resource the user is viewing is public or private.
+            //                  n/a is used for pages like discover and index
+            // page          => the name of the current page
+            // resource      => what resource the user is on. Ex node, preprint, registry.
+            // title         => the current route name
+            const {
+                authenticated,
+                isPublic,
+                resource,
+            } = config.metricsAdapters[0].dimensions;
+
             const page = document.location.pathname;
             const title = this.getWithDefault('currentRouteName', 'unknown');
+            const isAuthenticated = this.get('session.isAuthenticated');
+            const publicType = title.endsWith('content.index') ? 'public' : 'n/a';
+            const modelType = title.endsWith('content.index') ? 'preprint' : 'n/a';
 
-            this.get('metrics').trackPage({ page, title });
+            this.get('metrics').trackPage({
+                [authenticated]: isAuthenticated ? 'Logged in' : 'Logged out',
+                [isPublic]: publicType,
+                page,
+                [resource]: modelType,
+                title,
+            });
             this.set('theme.currentLocation', window.location.href);
         });
     },